Code refactorized.
