+++ app/Comment.php
@@ -10,7 +10,6 @@ use App\Repositories\CommentsRepository;
 use App\Resources\CommentResource;
 use Cog\Contracts\Love\Reactable\Models\Reactable as ReactableContract;
 use Cog\Laravel\Love\Reactable\Models\Traits\Reactable;
-use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 
@@ -50,25 +49,5 @@ class Comment extends Model implements ReactableContract, Resource
     public function updateViaRequest(Request $request)
     {
         return CommentsRepository::updateCommentViaRequest($request, $this);
-    }
-
-    /**
-     * Get lazy loaded relationships of Geofence.
-     *
-     * @return array
-     */
-    public function getLazyRelationshipsAttribute()
-    {
-        return ['commentable'];
-    }
-
-    /**
-     * @param Request $request
-     * @return LengthAwarePaginator
-     * @codeCoverageIgnore
-     */
-    public static function resolveModels(Request $request)
-    {
-        return self::with('user')->latest()->paginate(20);
     }
 }
