+++ app/Geofence.php
@@ -13,7 +13,6 @@ use App\Resources\GeofenceResource;
 use Cog\Contracts\Love\Reactable\Models\Reactable as ReactableContract;
 use Cog\Laravel\Love\Reactable\Models\Traits\Reactable;
 use Grimzy\LaravelMysqlSpatial\Eloquent\SpatialTrait;
-use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Http\Request;
 
@@ -103,30 +102,5 @@ class Geofence extends Model implements ReactableContract, Resource
     public function getRouteKeyName()
     {
         return 'uuid';
-    }
-
-    /**
-     * Get lazy loaded relationships of Geofence.
-     *
-     * @return array
-     */
-    public function getLazyRelationshipsAttribute()
-    {
-        return ['user', 'photo'];
-    }
-
-    /**
-     * @param Request $request
-     * @return LengthAwarePaginator
-     */
-    public static function resolveModels(Request $request)
-    {
-        return $request->user('api')->is_child ? Geofence::with('user', 'photo')->where(
-            'user_uuid',
-            $request->user('api')->user_uuid
-        )->orWhere('is_public', true)->latest()->paginate(20) : Geofence::with('user', 'photo')->where(
-            'user_uuid',
-            $request->user('api')->uuid
-        )->orWhere('is_public', true)->latest()->paginate(20);
     }
 }
