+++ app/Http/Controllers/Api/Resource/IndexController.php
@@ -36,6 +36,6 @@ class IndexController extends Controller
      */
     public function __invoke(IndexRequest $request)
     {
-        return response()->json(finder()->findClass(finder()->currentResource())::resolveModels($request), 200);
+        return response()->json(finder()->findClass(finder()->currentResource())->getInstances($request), 200);
     }
 }
