+++ app/Resources/ActionResource.php
@@ -2,12 +2,13 @@
 
 namespace App\Resources;
 
+use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Http\Request;
 
 trait ActionResource
 {
     /**
-     * Get available alert reactions.
+     * Get available action reactions.
      *
      * @return string
      * @codeCoverageIgnore TODO Support action reactions
@@ -19,12 +20,12 @@ trait ActionResource
     }
 
     /**
-     * Update alert validation rules.
+     * Update action validation rules.
      *
      * @param Request $request
      * @return array
      * @codeCoverageIgnore
-     * @justify Actions can't be updated, are system generated.
+     * @reason Actions can't be updated, are system generated.
      */
     public function updateRules(Request $request)
     {
@@ -32,15 +33,45 @@ trait ActionResource
     }
 
     /**
-     * Store alert validation rules.
+     * Store action validation rules.
      *
      * @param Request $request
      * @return array
      * @codeCoverageIgnore
-     * @justify Actions can't be created by request, are system generated.
+     * @reason Actions can't be created by request, are system generated.
      */
     public function storeRules(Request $request)
     {
         return [];
+    }
+
+    /**
+     * Get action relationships using lazy loading.
+     *
+     * @return array
+     */
+    public function getLazyRelationshipsAttribute()
+    {
+        return ['user'];
+    }
+
+    /**
+     * Get action instances.
+     *
+     * @param Request $request
+     * @return LengthAwarePaginator
+     */
+    public function getInstances(Request $request)
+    {
+        if (!$request->user('api')->is_child) {
+            return self::with('user')->whereIn('user_uuid', $request->user('api')
+                ->users
+                ->pluck('uuid')
+                ->push(
+                    $request->user('api')->uuid
+                )->toArray())->paginate(25);
+        } else {
+            return self::with('user')->where('user_uuid', $request->user('api')->uuid)->paginate(25);
+        }
     }
 }
