+++ app/Resources/GeofenceResource.php
@@ -4,6 +4,7 @@
 namespace App\Resources;
 
 use App\Rules\Coordinate;
+use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 
@@ -51,5 +52,32 @@ trait GeofenceResource
             'is_public' => 'required|boolean',
             'coordinates' => ['required', new Coordinate()],
         ];
+    }
+
+    /**
+     * Get geofence relationships using lazy loading.
+     *
+     * @return array
+     */
+    public function getLazyRelationshipsAttribute()
+    {
+        return ['user', 'photo'];
+    }
+
+    /**
+     * Get geofence instances.
+     *
+     * @param Request $request
+     * @return LengthAwarePaginator
+     */
+    public function getInstances(Request $request)
+    {
+        return $request->user('api')->is_child ? self::with('user', 'photo')->where(
+            'user_uuid',
+            $request->user('api')->user_uuid
+        )->orWhere('is_public', true)->latest()->paginate(20) : self::with('user', 'photo')->where(
+            'user_uuid',
+            $request->user('api')->uuid
+        )->orWhere('is_public', true)->latest()->paginate(20);
     }
 }
