+++ app/Resources/LocationResource.php
@@ -3,6 +3,7 @@
 namespace App\Resources;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Collection;
 
 trait LocationResource
 {
@@ -24,7 +25,7 @@ trait LocationResource
      * @param Request $request
      * @return array
      * @codeCoverageIgnore
-     * @justify Locations are device generated, can't be modified by user.
+     * @reason Locations are device generated, can't be modified by user.
      */
     public function updateRules(Request $request)
     {
@@ -37,10 +38,33 @@ trait LocationResource
      * @param Request $request
      * @return array
      * @codeCoverageIgnore
-     * @justify Locations are device generated, can't be predefined validation rules.
+     * @reason Locations are device generated, can't be predefined validation rules.
      */
     public function storeRules(Request $request)
     {
         return [];
+    }
+
+    /**
+     * Get location relationships using lazy loading.
+     *
+     * @return array
+     */
+    public function getLazyRelationshipsAttribute()
+    {
+        return ['trackable'];
+    }
+
+    /**
+     * Get location instances.
+     *
+     * @param Request $request
+     * @return Collection
+     * @codeCoverageIgnore
+     * @reason Locations uses custom LocationRepository query.
+     */
+    public function getInstances(Request $request)
+    {
+        return self::query()->get();
     }
 }
