+++ app/User.php
@@ -18,7 +18,6 @@ use Cog\Laravel\Love\Reactable\Models\Traits\Reactable;
 use Cog\Laravel\Love\Reacterable\Models\Traits\Reacterable;
 use Grimzy\LaravelMysqlSpatial\Eloquent\SpatialTrait;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
-use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Http\Request;
 use Illuminate\Notifications\Notifiable;
@@ -180,38 +179,5 @@ class User extends Authenticatable implements MustVerifyEmail, ReacterableContra
     public static function createViaRequest(Request $request)
     {
         return UsersRepository::createUserViaRequest($request);
-    }
-
-    /**
-     * Get lazy loaded relationships of Geofence.
-     *
-     * @return array
-     */
-    public function getLazyRelationshipsAttribute()
-    {
-        return ['photo', 'user'];
-    }
-
-    /**
-     * @param Request $request
-     * @return LengthAwarePaginator
-     */
-    public static function resolveModels(Request $request)
-    {
-        if (!is_null($request->user('api')->user_uuid)) {
-            return self::with('photo', 'user')->latest()->where([
-                ['user_uuid', $request->user('api')->user_uuid]
-            ])->orWhere([
-                ['uuid', $request->user('api')->user_uuid]
-            ])->paginate(20);
-        } elseif ($request->user('api')->is_admin) {
-            return self::with('photo', 'user')->latest()->paginate(20);
-        } else {
-            return self::with('photo', 'user')->latest()->where([
-                ['user_uuid', $request->user()->uuid]
-            ])->orWhere([
-                ['uuid', $request->user('api')->uuid]
-            ])->paginate(20);
-        }
     }
 }
